+++ src/card/sets/tests/runtime_support/nested_definitions.rs
@@ -365,6 +365,7 @@ pub(in super::super) fn assert_nested_definition_applied_effect(
         | AppliedEffectDef::DoesNotUntapDuringUntapStep
         | AppliedEffectDef::MayChooseNotToUntap
         | AppliedEffectDef::CannotBlock
+        | AppliedEffectDef::CannotAttack
         | AppliedEffectDef::CannotBeBlocked
         | AppliedEffectDef::CannotBeEnchanted
         | AppliedEffectDef::CannotBecomeEnchanted