+++ src/game/ability_layers.rs
@@ -284,6 +284,7 @@ impl Game {
             | AppliedEffectDef::DoesNotUntapDuringUntapStep
             | AppliedEffectDef::MayChooseNotToUntap
             | AppliedEffectDef::CannotBlock
+            | AppliedEffectDef::CannotAttack
             | AppliedEffectDef::CannotBeBlocked
             | AppliedEffectDef::CannotBeEnchanted
             | AppliedEffectDef::CannotBecomeEnchanted