+++ src/game/continuous_effects.rs
@@ -230,6 +230,7 @@ impl Game {
             | AppliedEffectDef::DoesNotUntapDuringUntapStep
             | AppliedEffectDef::MayChooseNotToUntap
             | AppliedEffectDef::CannotBlock
+            | AppliedEffectDef::CannotAttack
             | AppliedEffectDef::CannotBeBlocked
             | AppliedEffectDef::CannotBeEnchanted
             | AppliedEffectDef::CannotBecomeEnchanted