+++ src/game/tests.rs
@@ -375,6 +375,7 @@ mod attack_restrictions;
 mod aura_upkeep;
 mod banding;
 mod blocking_relationship;
+mod cannot_attack;
 mod cannot_block;
 mod casting_and_targets;
 mod casting_modes;